// src/types.ts
export type FineType = {
    amount: string
    amount_to_pay: string
    bill_at: string
    discount_at: null | any
    discount_size: null | any
    division_code: null | any
    division_name: string
    doc_number: string
    doc_type: string
    docs: any[]
    fssp_ip: null | any
    fssp_uin: null | any
    is_cached: false
    is_revoked: null | any
    kbk: string
    koap_code: string
    koap_text: null | any
    location: string
    name: string
    number: string
    pay_status: string
    payee_bank_account: string
    payee_bank_bik: string
    payee_bank_name: string
    payee_inn: string
    payee_kpp: string
    payee_oktmo: string
    payee_username: string
    payment_status: string
    pics: any[]
    quittance: string
    reg_cert: null | any
    violation_at: string
    violator_name: string
}